tipitaka/requirement.txt
scrapy 
python-dotenv
psycopg[binary,pool]